ADD : tetris architectures

+++ configurations/structure/reinforcement_learning_configuration.py
@@ -51,7 +51,7 @@ class ReinforcementLearningConfiguration:
         self.number_environment_runners: int = NotProvided
         self.number_environment_per_environment_runners: int = NotProvided
         self.number_cpus_per_environment_runners: int = NotProvided
-        self.number_gpus_per_environment_runners: int = NotProvided
+        self.number_gpus_per_environment_runners: float = NotProvided
         self.compress_observations: bool = NotProvided
 
         # Learners
